+++ src/lib/st-breadcrumbs/st-breadcrumbs.component.spec.ts
@@ -10,17 +10,22 @@
  */
 import { SimpleChange } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { Http } from '@angular/http';
-import { RouterTestingModule } from '@angular/router/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 
 import { StBreadCrumbsComponent } from './st-breadcrumbs.component';
+import { StBreadCrumbItem } from './st-breadcrumbs.interface';
+
+const item: StBreadCrumbItem = {
+   label: 'section4',
+   icon: 'icon-home'
+};
 
 describe('StBreadCrumbsComponent', () => {
    let component: StBreadCrumbsComponent;
    let fixture: ComponentFixture<StBreadCrumbsComponent>;
 
-   const menuMock: string[] = ['section0', 'section1', 'section2', 'section3', 'section4', 'section5', 'section6'];
+   const menuMock: StBreadCrumbItem[] = [{ icon: 'icon-home' }, { label: 'section1', icon: '' }, { label: 'section2', icon: '' },
+      { label: 'section3'}, { label: 'section4', icon: 'icon-home' }, { label: 'section5', icon: '' }, { label: 'section6', icon: '' }];
    const expectedIndexShowing5: number[] = [0, -1, 3, 4, 5, 6];
    const expectedIndexShowing10: number[] = [0, 1, 2, 3, 4, 5, 6];
 
@@ -78,14 +83,43 @@ describe('StBreadCrumbsComponent', () => {
       expect(component.select.emit).toHaveBeenCalledWith(2);
    });
 
-   it('Should get option when index is greater than -1', () => {
+   it('should show a option with label and icon', () => {
+      component.options[1] = item;
       fixture.detectChanges();
-      expect(component.getOption(3)).toEqual(menuMock[3]);
+      expect(component.hasLabel).toBeTruthy();
+      expect(component.hasIcon).toBeTruthy();
+      expect(component.getIcon(1)).toEqual('icon-home');
+      expect(component.getLabel(1)).toEqual('section4');
    });
 
+   it('Should get label when index is greater than -1', () => {
+      fixture.detectChanges();
+      expect(component.getLabel(3)).toEqual(menuMock[3].label);
+   });
+
+   it(`should not show anything if not have label`, () => {
+      let label = component.getLabel(0);
+      fixture.detectChanges();
+      expect(component.hasLabel(0)).toBeFalsy();
+   });
+
+   it('Should get icon when index is greater than -1', () => {
+      fixture.detectChanges();
+      expect(component.getIcon(4)).toEqual(menuMock[4].icon);
+   });
    it('Should get 3 dots when index is -1', () => {
       fixture.detectChanges();
-      expect(component.getOption(-1)).toEqual('...');
+      expect(component.getLabel(-1)).toEqual('...');
+   });
+   it('Should get icon empty when index is -1', () => {
+      fixture.detectChanges();
+      expect(component.getIcon(-1)).toEqual('');
+   });
+
+   it(`should not show anything if not have icon`, () => {
+      let icon = component.getIcon(3);
+      fixture.detectChanges();
+      expect(component.hasIcon(3)).toBeFalsy();
    });
 
    it('Should recalculate initPos and elements to show when changes', () => {
@@ -100,7 +134,7 @@ describe('StBreadCrumbsComponent', () => {
 
       expect(component.indexArray).toEqual([0, 1, 2, 3, 4, 5, 6]);
 
-      component.options = [...menuMock, 'section7'];
+      component.options = [...menuMock, {label: 'section7'}];
       component.ngOnChanges({ options: new SimpleChange(menuMock, component.options, true)});
       expect(component.indexArray).toEqual([0, -1, 2, 3, 4, 5, 6, 7]);
    });